+++ packages/web-components/keen-slider-navigation-arrows/src/components/keen-slider-navigation-arrows.ts
@@ -110,6 +110,7 @@ export const Component: CorrectComponentType<Props> = (props, { element }) => {
         setSlider(slider);
         slider.on("destroyed", () => setSlider());
       });
+
       onCleanup(removePlugin);
     },
   );