+++ shopify/packages/ajax-cart-components/src/utils/getValueFromPath.ts
@@ -2,12 +2,12 @@
 export type Path<T> = T extends readonly (infer U)[]
   ? `${number}` | `${number}.${Path<U>}` // Allows indexing into readonly arrays
   : T extends object
-  ? {
-    [K in keyof T & (string | number)]: T[K] extends object | any[]
-    ? `${K}` | `${K}.${Path<T[K]>}`
-    : `${K}`;
-  }[keyof T & (string | number)]
-  : never;
+    ? {
+        [K in keyof T & (string | number)]: T[K] extends object | any[]
+          ? `${K}` | `${K}.${Path<T[K]>}`
+          : `${K}`;
+      }[keyof T & (string | number)]
+    : never;
 
 // 🔥 Correctly Extracts the Value at the Given Path
 export type PathValue<
@@ -15,19 +15,19 @@ export type PathValue<
   P extends string,
 > = P extends `${infer Key}.${infer Rest}`
   ? Key extends keyof T
-  ? PathValue<T[Key], Rest>
-  : Key extends `${number}`
-  ? T extends readonly (infer U)[]
-  ? PathValue<U, Rest>
-  : undefined
-  : undefined
+    ? PathValue<T[Key], Rest>
+    : Key extends `${number}`
+      ? T extends readonly (infer U)[]
+        ? PathValue<U, Rest>
+        : undefined
+      : undefined
   : P extends keyof T
-  ? T[P]
-  : P extends `${number}`
-  ? T extends readonly (infer U)[]
-  ? U
-  : undefined
-  : undefined;
+    ? T[P]
+    : P extends `${number}`
+      ? T extends readonly (infer U)[]
+        ? U
+        : undefined
+      : undefined;
 
 export function getValueFromPath<T, P extends Path<T>>(
   obj: T,