+++ shopify/packages/ajax-cart/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @brytdesigns/shopify-ajax-cart
 
+## 1.0.46
+
+### Patch Changes
+
+- Updated dependencies
+  - @brytdesigns/shopify-ajax-cart-api@1.0.37
+
 ## 1.0.45
 
 ### Patch Changes